requirements.txt
pillow
stegano
pydub
streamlit